% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{apparelDemographics}
\alias{apparelDemographics}
\title{Customer Demographics for the Apparel Retailer Dataset}
\format{A \code{data.table} with 2201 rows and 2 variables:

\describe{
  \item{Id}{Customer Id}
  \item{Gender}{0=male, 1=female}
}}
\usage{
data("apparelDemographics")
}
\description{
This data contains additional demographic information on the customers in the
"apparelTrans" dataset. This information can be used as time-invariant covariates.
}
\keyword{datasets}
