% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{apparelDynCov}
\alias{apparelDynCov}
\title{Dynamic Covariates for the Apparel Retailer Dataset}
\format{
A data.table with 781355 rows and 5 variables
\describe{
  \item{Id}{Customer Id}
  \item{Cov.Date}{Date of contextual factor}
  \item{DM}{1 if customer was contacted with direct marketing in this time period}
  \item{High-Season}{1 if high season during this time period}
  \item{Gender}{0=male, 1=female}
}
}
\usage{
data("apparelDynCov")
}
\description{
This data contains direct marketing and seasonality information on customers in the "apparelTrans" dataset.
This information can be used as time-varying covariates.
}
\keyword{datasets}
