% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_interface_bgbb.R
\name{bgbb}
\alias{bgbb}
\alias{bgbb,clv.data-method}
\alias{bgbb,clv.data.static.covariates-method}
\alias{bgbb,clv.data.dynamic.covariates-method}
\title{BG/BB models - Work In Progress}
\usage{
\S4method{bgbb}{clv.data}(
  clv.data,
  start.params.model = c(),
  optimx.args = list(),
  verbose = TRUE,
  ...
)

\S4method{bgbb}{clv.data.static.covariates}(
  clv.data,
  start.params.model = c(),
  optimx.args = list(),
  verbose = TRUE,
  names.cov.life = c(),
  names.cov.trans = c(),
  start.params.life = c(),
  start.params.trans = c(),
  names.cov.constr = c(),
  start.params.constr = c(),
  reg.lambdas = c(),
  ...
)

\S4method{bgbb}{clv.data.dynamic.covariates}(
  clv.data,
  start.params.model = c(),
  optimx.args = list(),
  verbose = TRUE,
  names.cov.life = c(),
  names.cov.trans = c(),
  start.params.life = c(),
  start.params.trans = c(),
  names.cov.constr = c(),
  start.params.constr = c(),
  reg.lambdas = c(),
  ...
)
}
\description{
Fits BG/BB models on transactional data with static and without covariates.
}
