% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bgbb_nocov_LL_ind}
\alias{bgbb_nocov_LL_ind}
\alias{bgbb_nocov_LL_sum}
\title{BG/BB: LogLikelihood without covariates}
\usage{
bgbb_nocov_LL_ind(vLogparams, vX, vT_x, vT_cal, vN_cal)

bgbb_nocov_LL_sum(vLogparams, vX, vT_x, vT_cal, vN_cal)
}
\arguments{
\item{vLogparams}{vector with the BG/BB model parameters log scaled}

\item{vX}{Frequency vector of length n counting the numbers of purchases.}

\item{vT_x}{Recency vector of length n.}

\item{vT_cal}{Vector of length n indicating the total number of periods of observation.}
}
\value{
Returns the respective LogLikelihood value for the BG/BB model without covariates.
}
\description{
BG/BB without Covariates:

The function \code{bgbb_nocov_LL_ind} calculates the individual LogLikelihood
values for each customer for the given parameters.

The function \code{bgbb_nocov_LL_sum} calculates the LogLikelihood value summed
across customers for the given parameters.
}
\details{
\code{alpha, beta, gamma, delta} are the parameters used for estimation.\cr
TODO: add description of parameters
}
\references{
\url{https://github.com/cran/BTYD/}.
}
