% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bgbb_nocov_PAlive}
\alias{bgbb_nocov_PAlive}
\title{BG/BB: PAlive without covariates}
\usage{
bgbb_nocov_PAlive(alpha, beta, gamma, delta, vX, vT_x, vN_cal)
}
\arguments{
\item{alpha}{TODO: describe}

\item{beta}{TODO: describe}

\item{gamma}{TODO: describe}

\item{delta}{TODO: describe}

\item{vX}{Frequency vector of length n counting the numbers of purchases.}

\item{vT_x}{Recency vector of length n.}

\item{vT_cal}{Vector of length n indicating the total number of periods of observation.}
}
\value{
Returns a vector with the PAlive for each customer.
}
\description{
BG/BB without Covariates: Calculates the probability of a customer being alive
at the end of the calibration period.
}
\details{
\code{alpha} \code{beta} \code{gamma} \code{delta} estimated parameters in original scale
for the BG/BB model, namely (alpha, beta, gamma, delta).
}
