% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bgnbd_CET}
\alias{bgnbd_CET}
\alias{bgnbd_nocov_CET}
\title{BG/NBD: Conditional Expected Transactions without covariates}
\usage{
bgnbd_nocov_CET(r, alpha, a, b, nPeriods, vX, vT_x, vT_cal)
}
\arguments{
\item{r}{TODO: describe}

\item{alpha}{TODO: describe}

\item{a}{TODO: describe}

\item{b}{TODO: describe}

\item{nPeriods}{time prediction time frame}

\item{vX}{Frequency vector of length n counting the numbers of purchases.}

\item{vT_x}{Recency vector of length n.}

\item{vT_cal}{Vector of length n indicating the total number of periods of observation.}
}
\value{
Returns a vector containing the conditional expected transactions for the existing
customers in the BG/NBD model.
}
\description{
Calculates the expected number of transactions in a given time period based
on a customer's past transaction behavior and the BG/NBD model parameters.
}
\details{
\code{r} \code{alpha} \code{a} \code{b} estimated parameters in original scale
for the BG/NBD model, namely (r, alpha, a, b).
r and alpha: TODO: description.
a and b: TODO: description
}
