% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bgnbd_LL}
\alias{bgnbd_LL}
\alias{bgnbd_nocov_LL_ind}
\alias{bgnbd_nocov_LL_sum}
\alias{bgnbd_staticcov_LL_ind}
\alias{bgnbd_staticcov_LL_sum}
\title{BG/NBD: Log-Likelihood functions}
\usage{
bgnbd_nocov_LL_ind(vLogparams, vX, vT_x, vT_cal)

bgnbd_nocov_LL_sum(vLogparams, vX, vT_x, vT_cal)

bgnbd_staticcov_LL_ind(vLogparams, vX, vT_x, vT_cal, mCov_life, mCov_trans)

bgnbd_staticcov_LL_sum(vLogparams, vX, vT_x, vT_cal, mCov_life, mCov_trans)
}
\arguments{
\item{vLogparams}{vector with the BG/NBD model parameters log scaled}

\item{vX}{Frequency vector of length n counting the numbers of purchases.}

\item{vT_x}{Recency vector of length n.}

\item{vT_cal}{Vector of length n indicating the total number of periods of observation.}

\item{mCov_life}{Matrix containing the covariates data affecting the lifetime process. One column for each covariate.}

\item{mCov_trans}{Matrix containing the covariates data affecting the transaction process. One column for each covariate.}
}
\value{
Returns the respective Log-Likelihood value(s) for the BG/NBD model
 with or without covariates.
}
\description{
Calculates the Log-Likelihood values for the BG/NBD model with and without covariates.

The function \code{bgnbd_nocov_LL_ind} calculates the individual LogLikelihood
values for each customer for the given parameters.

The function \code{bgnbd_nocov_LL_sum} calculates the LogLikelihood value summed
across customers for the given parameters.

The function \code{bgnbd_staticcov_LL_ind} calculates the individual LogLikelihood
values for each customer for the given parameters and covariates.

The function \code{bgnbd_staticcov_LL_sum} calculates the individual LogLikelihood values summed
across customers.
}
\details{
\code{vLogparams} is vector with the BG/NBD model parameters at log scale,
followed by the parameters for the lifetime covariate at original scale and then
followed by the parameters for the transaction covariate at original scale.

Model parameters for the BG/NBD model are \code{r, alpha, a, and b}. \cr
\code{r}: shape parameter of the Gamma distribution of the purchase process. \cr
\code{alpha}: scale parameter of the Gamma distribution of the purchase process. \cr
\code{a}: shape parameter of the Beta distribution of the dropout process.\cr
\code{b}: shape parameter of the Beta distribution of the dropout process.

\code{mCov_trans} is a matrix containing the covariates data of
the time-invariant covariates that affect the transaction process.
Each column represents a different covariate. For every column a gamma parameter
needs to added to \code{vLogparams} at the respective position.

\code{mCov_life} is a matrix containing the covariates data of
the time-invariant covariates that affect the lifetime process.
Each column represents a different covariate. For every column a gamma parameter
needs to added to \code{vLogparams} at the respective position.
}
\references{
Fader PS, Hardie BGS, Lee, KL (2005). \dQuote{\dQuote{Counting Your Customers} the Easy Way:
An Alternative to the Pareto/NBD Model} Marketing Science, 24(2), 275–284.

Fader PS, Hardie BGS (2013). \dQuote{Overcoming the BG/NBD Model’s #NUM! Error Problem}
URL \url{http://brucehardie.com/notes/027/bgnbd_num_error.pdf}.

Fader PS, Hardie BGS (2007). \dQuote{Incorporating time-invariant covariates into the
Pareto/NBD and BG/NBD models.}
URL \url{http://www.brucehardie.com/notes/019/time_invariant_covariates.pdf}.
}
