% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bgnbd_nocov_LL_sum}
\alias{bgnbd_nocov_LL_sum}
\title{BG/NBD: LogLikelihood without covariates}
\usage{
bgnbd_nocov_LL_sum(vLogparams, vX, vT_x, vT_cal)
}
\arguments{
\item{vLogparams}{vector with the Pareto/NBD model parameters log scaled}

\item{vX}{Frequency vector of length n counting the numbers of purchases.}

\item{vT_x}{Recency vector of length n.}

\item{vT_cal}{Vector of length n indicating the total number of periods of observation.}
}
\value{
Returns the respective LogLikelihood value for the BG/NBD model without covariates.
}
\description{
BG/NBD without Covariates:

The function \code{bgnbd_nocov_LL_ind} calculates the individual LogLikelihood
values for each customer for the given parameters.

The function \code{bgnbd_nocov_LL_sum} calculates the LogLikelihood value summed
across customers for the given parameters.
}
\details{
\code{r, alpha, a, b} are the parameters used for estimation.\cr
TODO: add description of parameters
}
\references{
\url{https://github.com/cran/BTYD/}.
}
