% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bgnbd_PAlive}
\alias{bgnbd_PAlive}
\alias{bgnbd_palive}
\title{BG/NBD: PAlive without covariates}
\usage{
bgnbd_palive(r, alpha, a, b, vX, vT_x, vT_cal)
}
\arguments{
\item{r}{TODO: describe}

\item{alpha}{TODO: describe}

\item{a}{TODO: describe}

\item{b}{TODO: describe}

\item{vX}{Frequency vector of length n counting the numbers of purchases.}

\item{vT_x}{Recency vector of length n.}

\item{vT_cal}{Vector of length n indicating the total number of periods of observation.}

\item{nPeriods}{time prediction time frame}
}
\value{
Returns a vector with the PAlive for each customer.
}
\description{
BG/NBD without Covariates: Calculates the probability of a customer being alive
at the end of the calibration period.
}
\details{
\code{r} \code{alpha} \code{a} \code{b} estimated parameters in original scale
for the BG/NBD model, namely (r, alpha, a, b).
r and alpha: TODO: description.
a and b: TODO: description
}
