% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bgnbd_staticcov_CET}
\alias{bgnbd_staticcov_CET}
\title{BG/NBD: Conditional Expected Transactions with static covariates}
\usage{
bgnbd_staticcov_CET(
  r,
  alpha,
  a,
  b,
  nPeriods,
  vX,
  vT_x,
  vT_cal,
  vCovParams_trans,
  vCovParams_life,
  mCov_trans,
  mCov_life
)
}
\arguments{
\item{vX}{Frequency vector of length n counting the numbers of purchases.}

\item{vT_x}{Recency vector of length n.}

\item{vT_cal}{Vector of length n indicating the total number of periods of observation.}

\item{vCovParams_trans}{Vector of estimated parameters for the transaction covariates.}

\item{vCovParams_life}{Vector of estimated parameters for the dropout covariates.}

\item{mCov_trans}{Matrix containing the covariates data affecting the transaction process. One column for each covariate.}

\item{mCov_life}{Matrix containing the covariates data affecting the lifetime process. One column for each covariate.}

\item{vEstimated_params}{Vector with the Pareto/NBD model parameters at original scale.}

\item{dPrediction_period}{time prediction time frame}
}
\value{
Returns a vector with the conditional expected transactions for the existing
customers in the BG/NBD model.
}
\description{
Calculates the expected number of transactions in a given time period based
on a customer's past transaction behavior and the BG/NBD model parameters.
}
\details{
\code{r} r: TODO

\code{r} alpha: TODO

\code{r} a: TODO

\code{r} b: TODO

\code{mCov_trans} is a matrix containing the covariates data of
the time-invariant covariates that affect the transaction process.
Each column represents a different covariate. For every column a gamma parameter
needs to added to \code{vCovParams_trans} at the respective position.

\code{mCov_life} is a matrix containing the covariates data of
the time-invariant covariates that affect the lifetime process.
Each column represents a different covariate. For every column a gamma parameter
needs to added to \code{vCovParams_life} at the respective position.
}
\references{
TODO
}
