% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bgnbd_staticcov_LL_ind}
\alias{bgnbd_staticcov_LL_ind}
\alias{bgnbd_staticcov_LL_sum}
\title{BG/NBD: LogLikelihood with static covariates}
\usage{
bgnbd_staticcov_LL_ind(vLogparams, vX, vT_x, vT_cal, mCov_life, mCov_trans)

bgnbd_staticcov_LL_sum(vLogparams, vX, vT_x, vT_cal, mCov_life, mCov_trans)
}
\arguments{
\item{vLogparams}{vector with the parameters for the BG/NBD model and the static covariates. See Details.}

\item{vX}{Frequency vector of length n counting the numbers of purchases.}

\item{vT_x}{Recency vector of length n.}

\item{vT_cal}{Vector of length n indicating the total number of periods of observation.}

\item{mCov_life}{Matrix containing the covariates data affecting the lifetime process. One column for each covariate.}

\item{mCov_trans}{Matrix containing the covariates data affecting the transaction process. One column for each covariate.}
}
\value{
Returns the respective LogLikelihood value for the BG/NBD model with static covariates.
}
\description{
BG/NBD with Static Covariates:

The function \code{bgnbd_staticcov_LL_ind} calculates the individual LogLikelihood
values for each customer for the given parameters and covariates.

The function \code{bgnbd_staticcov_LL_sum} calculates the individual LogLikelihood values summed
across customers.
}
\details{
\code{vLogparams} is vector with the BG/NBD model parameters at log scale,
followed by the parameters for the lifetime covariate at original scale and then
followed by the parameters for the transaction covariate at original scale

\code{mCov_life} is a matrix containing the covariates data of
the time-invariant covariates that affect the lifetime process.
Each column represents a different covariate. For every column, a gamma parameter
needs to added to \code{vParams} at the respective position.

\code{mCov_trans} is a matrix containing the covariates data of
the time-invariant covariates that affect the transaction process.
Each column represents a different covariate. For every column, a gamma parameter
needs to added to \code{vParams} at the respective position.
}
\references{
TODO
}
