% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdnow}
\alias{cdnow}
\title{CDNOW dataset}
\format{
A \code{data.table} with 6696 rows and 4 variables:
\describe{
  \item{\code{Id}}{Customer Id}
  \item{\code{Date}}{Date of purchase}
  \item{\code{CDs}}{Amount of CDs purchased}
  \item{\code{Price}}{Price of purchase}
}
}
\usage{
data("cdnow")
}
\description{
A dataset containing the entire purchase history up to the end of June 1998
of the cohort of 23,570 individuals who made their first-ever purchase at CDNOW in the first quarter
of 1997.
}
\references{
Fader, Peter S. and Bruce G.,S. Hardie, (2001), "Forecasting Repeat Sales at CDNOW:
A Case Study," Interfaces, 31 (May-June), Part 2 of 2, p94-107.
}
\keyword{datasets}
