% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_clv_bgnbd_staticcov.R
\docType{class}
\name{clv.bgnbd.static.cov-class}
\alias{clv.bgnbd.static.cov-class}
\title{Result of fitting the BG/NBD model with static covariates}
\description{
Output from fitting the BG/NBD model on data with static covariates. It constitutes the
estimation result and is returned to the user to use it as input to other methods such as
to make predictions or plot the unconditional expectation.

Inherits from \linkS4class{clv.fitted.transactions.static.cov} in order to execute all steps required for fitting a model
with static covariates and it contains an instance of class \linkS4class{clv.model.bgnbd.static.cov} which
provides the required BG/NBD (static covariates) specific functionalities.
}
\section{Slots}{

\describe{
\item{\code{cbs}}{Single \code{data.table} that is the Customer-By-Sufficiency matrix with information about
each customers' recency, frequency, and total time observed in number of time-units
and as defined by the BG/NBD model.}
}}

\seealso{
\linkS4class{clv.fitted.transactions.static.cov}, \linkS4class{clv.model.bgnbd.static.cov}, \linkS4class{clv.bgnbd}
}
\keyword{internal}
