% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_clv_data.R
\name{clv.data.mean.interpurchase.times}
\alias{clv.data.mean.interpurchase.times}
\title{Interpurchase time, for repeaters only
  Time between consecutive purchases of each customer - convert to intervals then time units
  If zero-repeaters (only 1 trans) set NA to ignore it in mean / sd calculations}
\usage{
clv.data.mean.interpurchase.times(clv.data, dt.transactions)
}
\description{
Interpurchase time, for repeaters only
  Time between consecutive purchases of each customer - convert to intervals then time units
  If zero-repeaters (only 1 trans) set NA to ignore it in mean / sd calculations
}
