% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_clv_data_staticcovariates.R,
%   R/f_s3generics_clvdata_staticcov.R
\docType{class}
\name{clv.data.static.covariates-class}
\alias{clv.data.static.covariates-class}
\alias{show,clv.data.static.covariates-method}
\title{Transactional and static covariates data to fit CLV models}
\usage{
\S4method{show}{clv.data.static.covariates}(object)
}
\description{
Extends the class \linkS4class{clv.data} and adds slots to store data and names of
static covariates for both processes.
An object of this class then serves as input to fit models with static covariates.
}
\section{Slots}{

\describe{
\item{\code{data.cov.life}}{Single \code{data.table} with all static covariate data for the lifetime process}

\item{\code{data.cov.trans}}{Single \code{data.table} with all static covariate data for the transaction process}

\item{\code{names.cov.data.life}}{Character vector with names of the static lifetime covariates.}

\item{\code{names.cov.data.trans}}{Character vector with names of the static transaction covariates.}
}}

\seealso{
Definition of the parent class \linkS4class{clv.data}.

For fitting covariate models: \code{\link[CLVTools:pnbd]{pnbd}}
}
\keyword{internal}
