% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_clv_fitted_dynamiccov.R
\docType{class}
\name{clv.fitted.dynamic.cov-class}
\alias{clv.fitted.dynamic.cov-class}
\title{Fitted CLV Model with Dynamic covariates}
\description{
Extends the class \code{\link[CLVTools:clv.fitted.static.cov-class]{clv.fitted.static.cov}} but adds no
additional slots to it. The purpose of this class rather is to perform steps during the fitting
process that are specific to dynamic covariates models.
}
\seealso{
\code{\link[CLVTools:clv.fitted-class]{clv.fitted}}, \code{\link[CLVTools:clv.fitted.static.cov-class]{clv.fitted.static.cov}}
}
\keyword{internal}
