% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_clv_fitted_spending.R
\docType{class}
\name{clv.fitted.spending-class}
\alias{clv.fitted.spending-class}
\title{Fitted Spending Model}
\description{
Extends the class \code{clv.fitted} with slots for spending models and performs steps during the
estimation, prediction and plotting process that are specific to all spending models.
}
\section{Slots}{

\describe{
\item{\code{estimation.removed.first.transaction}}{Single boolean whether every customers' first transaction was removed when counting transactions.}
}}

\seealso{
Definition of the parent class \linkS4class{clv.fitted}

For transaction models \linkS4class{clv.fitted.transactions}
}
\keyword{internal}
