% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_clv_fitted_transactions.R
\docType{class}
\name{clv.fitted.transactions-class}
\alias{clv.fitted.transactions-class}
\title{Fitted CLV Transaction Model without covariates}
\description{
Fitted CLV Transaction Model without covariates
}
\keyword{internal}
