% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_clv_fitted_transactions_dynamiccov.R
\docType{class}
\name{clv.fitted.transactions.dynamic.cov-class}
\alias{clv.fitted.transactions.dynamic.cov-class}
\title{Fitted CLV Model with Dynamic covariates}
\description{
Extends the class \linkS4class{clv.fitted.transactions.static.cov} but adds no
additional slots to it. The purpose of this class rather is to perform steps during the fitting
process that are specific to dynamic covariates models.
}
\seealso{
\linkS4class{clv.fitted}, \linkS4class{clv.fitted.transactions.static.cov}
}
\keyword{internal}
