% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_clv_gg.R
\docType{class}
\name{clv.gg-class}
\alias{clv.gg-class}
\title{Result of fitting the Gamma/Gamma model without covariates}
\description{
Output from fitting the Gamma/Gamma model on data without covariates. It constitutes the estimation
result and is returned to the user to use it as input to other methods such as to make
predictions or plot the unconditional expectation.

Inherits from \code{clv.fitted} in order to execute all steps required for fitting a model
without covariates and it contains an instance of class \code{clv.model.gg} which
provides the required Gamma/Gamma (no covariates) specific functionalities.
}
\section{Slots}{

\describe{
\item{\code{cbs}}{Single \code{data.table} that is the Customer-By-Sufficiency matrix with information about
each customer as defined by the Gamma/Gamma model.}
}}

\seealso{
\link{clv.fitted-class}, \link{clv.model.gg-class}
}
\keyword{internal}
