% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_clv_model_nocorrelation.R
\docType{class}
\name{clv.model.no.correlation-class}
\alias{clv.model.no.correlation-class}
\title{CLV Model without support for life-trans correlation}
\description{
This class serves as a parent class to other clv.model classes that cannot be fit with correlation between the
lifetime and transaction process.
}
\seealso{
Parent class \linkS4class{clv.model}

CLV model subclasses  \linkS4class{clv.model.bgnbd.no.cov}, \linkS4class{clv.model.ggomnbd.no.cov}
}
\keyword{internal}
