% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_clv_model_pnbd_dynamiccov.R
\docType{class}
\name{clv.model.pnbd.dynamic.cov-class}
\alias{clv.model.pnbd.dynamic.cov-class}
\title{CLV Model functionality for PNBD with dynamic covariates}
\description{
This class implements the functionalities and model-specific steps which are required
to fit the Pareto/NBD model with dynamic covariates.
}
\seealso{
Other clv model classes \link{clv.model-class}, \link{clv.model.pnbd.no.cov-class}, \link{clv.model.pnbd.static.cov-class}

Classes using its instance: \link{clv.fitted.dynamic.cov-class},
}
\keyword{internal}
