% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_clv_model_withcorrelation.R
\docType{class}
\name{clv.model.with.correlation-class}
\alias{clv.model.with.correlation-class}
\title{CLV Model providing life-trans correlation related functionalities}
\description{
This class serves as a parent class to other clv.model classes and provides the functionality needed
to fit a model with correlation between the lifetime and transaction process.
}
\section{Slots}{

\describe{
\item{\code{estimation.used.correlation}}{Single boolean whether the correlation was estimated.}

\item{\code{name.prefixed.cor.param.m}}{Single character vector of the internal name used for the correlation parameter during optimization.}

\item{\code{name.correlation.cor}}{Single character vector of the external name used for the correlation parameter.}
}}

\seealso{
Parent class \linkS4class{clv.model}

CLV model subclasses  \linkS4class{clv.model.pnbd.no.cov}, \linkS4class{clv.model.pnbd.static.cov}, \linkS4class{clv.model.pnbd.dynamic.cov}
}
\keyword{internal}
