% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_clv_pnbd.R
\docType{class}
\name{clv.pnbd-class}
\alias{clv.pnbd-class}
\title{Result of fitting the Pareto/NBD model without covariates}
\description{
Output from fitting the Pareto/NBD model on data without covariates. It constitutes the estimation
result and is returned to the user to use it as input to other methods such as to make
predictions or plot the unconditional expectation.

Inherits from \linkS4class{clv.fitted.transactions} in order to execute all steps required for fitting a model
without covariates and it contains an instance of class \linkS4class{clv.model.pnbd.no.cov} which
provides the required Pareto/NBD (no covariates) specific functionalities.
}
\section{Slots}{

\describe{
\item{\code{cbs}}{Single \code{data.table} that is the Customer-By-Sufficiency matrix with information about
each customers' recency, frequency, and total time observed in number of time-units
and as defined by the Pareto/NBD model.}
}}

\seealso{
\linkS4class{clv.fitted.transactions}, \linkS4class{clv.model.pnbd.no.cov}, \linkS4class{clv.pnbd.static.cov}, \linkS4class{clv.pnbd.dynamic.cov}
}
\keyword{internal}
