% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_clv_pnbd_dynamiccov.R
\docType{class}
\name{clv.pnbd.dynamic.cov-class}
\alias{clv.pnbd.dynamic.cov-class}
\title{Result of fitting the Pareto/NBD model with dynamic covariates}
\description{
Output from fitting the Pareto/NBD model on data with dynamic covariates. It constitutes the
estimation result and is returned to the user to use it as input to other methods such as
to make predictions or plot the unconditional expectation.

Inherits from \code{clv.fitted.transactions.dynamic.cov} in order to execute all steps required for fitting a model
with dynamic covariates and it contains an instance of class \code{clv.model.pnbd.dynamic.cov} which
provides the required Pareto/NBD (dynamic covariates) specific functionalities.
}
\section{Slots}{

\describe{
\item{\code{cbs}}{Single \code{data.table} that is the Customer-By-Sufficiency matrix with information about
each customers' recency, frequency, and total time observed in number of time-units
and as defined by the Pareto/NBD model.}
}}

\seealso{
\linkS4class{clv.fitted.transactions.dynamic.cov}, \linkS4class{clv.model.pnbd.dynamic.cov}, \linkS4class{clv.pnbd}, \linkS4class{clv.pnbd.static.cov}
}
\keyword{internal}
