% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_clv_time_days.R
\docType{class}
\name{clv.time.days-class}
\alias{clv.time.days-class}
\title{Time unit representing a single Day}
\description{
Time unit representing a single Day
}
\seealso{
\code{\link[CLVTools:clv.time]{clv.time}}

\code{\link[CLVTools:clv.time.date]{clv.time.date}}
}
\keyword{internal}
