% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_clv_time_weeks.R
\docType{class}
\name{clv.time.weeks-class}
\alias{clv.time.weeks-class}
\title{Time unit representing a single Week}
\description{
Inherits from \code{clv.time.date} and therefore stores all timepoints
with instances of class \code{Date}. Hence, no temporal resolution finer than 1 day is
possible with this time unit.
}
\keyword{internal}
