% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_s3generics_clvdata_plot.R
\name{density.clv.data}
\alias{density.clv.data}
\title{Plot Density of Spending / Transaction Values}
\usage{
\method{density}{clv.data}(x, mean.spending = TRUE, color = "black", geom = "line", n = 256, ...)
}
\arguments{
\item{x}{object of class \code{clv.data}}

\item{mean.spending}{xxxxxxxx}

\item{color}{xxxxxxxx}

\item{geom}{The geometric object of ggplo2 to display the data. Forwarded to \link[ggplot2:stat_density]{ggplot2::stat_density}.}

\item{n}{Number of points at which the empirical and model density are calculated. Should be a power of two.}

\item{...}{Forwarded to \link[ggplot2:stat_density]{ggplot2::stat_density}.}
}
\description{
Density plot of average spending per.
Note that this includes every transaction and not only repeat-transactions.
}
\examples{


}
