% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{donationTrans}
\alias{donationTrans}
\title{Discrete donation dataset}
\format{
A \code{data.table} with 52432 rows and 2 variables:
\describe{
  \item{\code{Id}}{Customer Id}
  \item{\code{Date}}{Date of donnations}
}
}
\usage{
data("donationTrans")
}
\description{
A simulated dataset containing a simulated donation history of 10000 individuals from 1970
up to 1983. The individuals have a yearly opportunity to donate.
}
\references{
Dziurzynski, Lukasz, Wadsworth, Edward and McCarthy, Daniel, (2014), "BTYD: Implementing Buy 'Til You Die Models"
R Package version 2.4, https://CRAN.R-project.org/package=BTYD.
}
\keyword{datasets}
