% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{factorial_C}
\alias{factorial_C}
\title{Compute factorial using gamma function}
\arguments{
\item{n}{Integer value for which to calculate factorial}
}
\value{
The factorial of n
}
\description{
Calculates factorial using the gamma function for numerical stability
}
\keyword{internal}
