% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_s3generics_clvfitted.R
\name{fitted.clv.fitted}
\alias{fitted.clv.fitted}
\title{Extract Unconditional Expectation}
\usage{
\method{fitted}{clv.fitted}(object, prediction.end = NULL, verbose = FALSE, ...)
}
\arguments{
\item{object}{A fitted clv model for which the unconditional expectation is desired.}

\item{prediction.end}{Until what point in time to predict. This can be the number of periods (numeric) or a form of date/time object. See details.}

\item{verbose}{Show details about the running of the function.}

\item{...}{Ignored}
}
\description{
Extract the unconditional expectation (future transactions unconditional on beein "alive") from a fitted clv model.
}
\details{
\code{prediction.end} is either a point in time (of class \code{Date}, \code{POSIXct}, or \code{character}) or the number of periods
that indicates until when to calculate the unconditional expectation.
If \code{prediction.end} is of class character, the date/time format set when creating the data object is used for parsing.
If \code{prediction.end} is the number of periods, the end of the fitting period serves as the reference point from which periods are counted. Only full periods may be specified.
If \code{prediction.end} is omitted or NULL, it defaults to the end of the holdout period.
}
