% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ggomnbd_nocov_CET}
\alias{ggomnbd_nocov_CET}
\title{GGompertz/NBD: Conditional Expected Transactions without Covariates}
\usage{
ggomnbd_nocov_CET(vEstimated_params, dPrediction_period, vX, vT_x, vT_cal)
}
\arguments{
\item{vEstimated_params}{Vector with the GGompertz/NBD model parameters at original scale.}

\item{dPrediction_period}{time prediction time frame}

\item{vX}{Frequency vector of length n counting the numbers of purchases.}

\item{vT_x}{Recency vector of length n.}

\item{vT_cal}{Vector of length n indicating the total number of periods of observation.}
}
\value{
Returns a vector with the conditional expected transactions for the existing
customers in the GGompertz/NBD model.
}
\description{
Calculates the expected number of transactions in a given time period based
on a customers past transaction behavior and the GGompertz/NBD model parameters.
}
\details{
\code{vEstimated_params}
s: shape parameter of the Gamma distribution for the lifetime process.\cr
The smaller s, the stronger the heterogeneity of customer lifetimes. \cr
beta: scale parameter for the Gamma distribution for the lifetime process. \cr
b: scale parameter of the Gompertz distribution (constant across customers). \cr
r: shape parameter of the Gamma distribution of the purchase process.
The smaller r, the stronger the heterogeneity of the pruchase process.\cr
alpha: scale parameter of the Gamma distribution of the purchase process.

\code{dPrediction_period} is the duration over which the prediction is made.
 Usually this is the duration of the holdout period.
}
