% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ggomnbd_nocov_LL_ind}
\alias{ggomnbd_nocov_LL_ind}
\alias{ggomnbd_nocov_LL_sum}
\title{GGompertz/NBD: LogLikelihood without covariates}
\usage{
ggomnbd_nocov_LL_ind(vLogparams, vX, vT_x, vT_cal)

ggomnbd_nocov_LL_sum(vLogparams, vX, vT_x, vT_cal)
}
\arguments{
\item{vLogparams}{vector with the GGompertz/NBD model parameters at log scale}

\item{vX}{Frequency vector of length n counting the numbers of purchases.}

\item{vT_x}{Recency vector of length n.}

\item{vT_cal}{Vector of length n indicating the total number of periods of observation.}
}
\value{
Returns the respective LogLikelihood value for the GGompertz/NBD Model without covariates.
}
\description{
The function \code{ggomnbd_nocov_LL_ind} calculates the individual LogLikelihood
values for each customer for the given parameters.

The function \code{ggomnbd_nocov_LL_sum} calculates the LogLikelihood value summed
across customers for the given parameters.
}
\details{
\code{r, alpha_0, b, s, beta_0} are the log()-ed model parameters used for
estimation, in this order.\cr
\code{s}: shape parameter of the Gamma distribution for the lifetime process.
The smaller \code{s}, the stronger the heterogeneity of customer lifetimes.\cr
\code{beta}: scale parameter for the Gamma distribution for the lifetime process.\cr
\code{b:} scale parameter of the Gompertz distribution (constant across customers).\cr
\code{r:} shape parameter of the Gamma distribution of the purchase process.
The smaller \code{r}, the stronger the heterogeneity of the pruchase process.\cr
\code{alpha}: scale parameter of the Gamma distribution of the purchase process.\cr


Ideally, the starting parameters for r and s represent your best guess
concerning the heterogeneity of customers in their buy and die rate.
}
