% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ggomnbd_nocov_PAlive}
\alias{ggomnbd_nocov_PAlive}
\title{GGompertz/NBD: PAlive without Covariates}
\usage{
ggomnbd_nocov_PAlive(vEstimated_params, vX, vT_x, vT_cal)
}
\arguments{
\item{vEstimated_params}{Vector with the GGompertz/NBD model parameters at original scale.}

\item{vX}{Frequency vector of length n counting the numbers of purchases.}

\item{vT_x}{Recency vector of length n.}

\item{vT_cal}{Vector of length n indicating the total number of periods of observation.}
}
\value{
Returns a vector containing the PAlive for each customer.
}
\description{
Calculates the probability of a customer being alive (PAlive) at the
end of the calibration period in the GGompertz/NBD model
}
\details{
\code{r, alpha_0, b, s, beta_0} are the log()-ed model parameters used for
estimation, in this order.\cr
\code{s}: shape parameter of the Gamma distribution for the lifetime process. The smaller \code{s}, the stronger the heterogeneity of customer lifetimes.\cr
\code{beta}: scale parameter for the Gamma distribution for the lifetime process.\cr
\code{b:} scale parameter of the Gompertz distribution (constant across customers).\cr
\code{r:} shape parameter of the Gamma distribution of the purchase process. The smaller \code{r}, the stronger the heterogeneity of the pruchase process.\cr
\code{alpha}: scale parameter of the Gamma distribution of the purchase process.\cr
}
