% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ggomnbd_staticcov_LL_sum}
\alias{ggomnbd_staticcov_LL_sum}
\title{GGompertz/NBD: LogLikelihood with static covariates}
\usage{
ggomnbd_staticcov_LL_sum(vParams, vX, vT_x, vT_cal, mCov_life, mCov_trans)
}
\arguments{
\item{vParams}{vector with the parameters for the GGompertz/NBD model and for the static
covariates. See Details.}

\item{vX}{Frequency vector of length n counting the numbers of purchases.}

\item{vT_x}{Recency vector of length n.}

\item{vT_cal}{Vector of length n indicating the total number of periods of observation.}

\item{mCov_life}{Matrix containing the covariates data affecting the lifetime process. One column for each covariate.}

\item{mCov_trans}{Matrix containing the covariates data affecting the transaction process. One column for each covariate.}
}
\value{
Returns the respective LogLikelihood value for the GGompertz/NBD model with static covariates.
}
\description{
The function \code{ggomnbd_staticcov_LL_ind} calculates the individual LogLikelihood
values for each customer for the given parameters.

The function \code{ggomnbd_staticcov_LL_sum} calculates the LogLikelihood value summed
across customers for the given parameters.

\code{mCov_life} is a matrix containing the covariates data of
the time-invariant covariates that affect the lifetime process.
Each column represents a different covariate. For every column, a gamma parameter
needs to added to \code{vParams} at the respective position.
}
\details{
\code{vParams} is vector with the GGompertz/NBD model parameters at log scale
(\code{r, alpha_0, b, s, beta_0}), followed by the parameters for the lifetime
covariate at original scale (\code{mCov_life}) and then followed by the parameters
for the transaction covariate at original scale \code{mCov_trans}.
\code{r, alpha_0, b, s, beta_0} are the log()-ed model parameters used for
estimation, in this order.\cr
\code{s}: shape parameter of the Gamma distribution for the lifetime process.
The smaller \code{s}, the stronger the heterogeneity of customer lifetimes.\cr
\code{beta}: scale parameter for the Gamma distribution for the lifetime process.\cr
\code{b:} scale parameter of the Gompertz distribution (constant across customers).\cr
\code{r:} shape parameter of the Gamma distribution of the purchase process.
The smaller \code{r}, the stronger the heterogeneity of the pruchase process.\cr
\code{alpha}: scale parameter of the Gamma distribution of the purchase process.\cr
\code{mCov_life}: parameters for the covariates affecting the lifetime process.\cr
\code{mCov_trans}: parameters for the covariates affecting the transaction process.
}
