% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ggomnbd_staticcov_PAlive}
\alias{ggomnbd_staticcov_PAlive}
\title{GGompertz/NBD: PAlive with Static Covariates}
\usage{
ggomnbd_staticcov_PAlive(
  vEstimated_params,
  vX,
  vT_x,
  vT_cal,
  vCovParams_trans,
  vCovParams_life,
  mCov_life,
  mCov_trans
)
}
\arguments{
\item{vEstimated_params}{Vector with the GGompertz/NBD model parameters at original scale.}

\item{vX}{Frequency vector of length n counting the numbers of purchases.}

\item{vT_x}{Recency vector of length n.}

\item{vT_cal}{Vector of length n indicating the total number of periods of observation.}

\item{vCovParams_trans}{Vector of estimated parameters for the transaction covariates.}

\item{vCovParams_life}{Vector of estimated parameters for the dropout covariates.}

\item{mCov_life}{Matrix containing the covariates data affecting the lifetime process. One column for each covariate.}

\item{mCov_trans}{Matrix containing the covariates data affecting the transaction process. One column for each covariate.}
}
\value{
Returns a vector containing the PAlive for each customer.
}
\description{
Calculates the probability of a customer being alive (PAlive) at the
end of the calibration period.
}
\details{
\code{vEstimated_params} is vector with the five estimated model parameters of the
GGompertz/NBD model parameters at original scale (\code{r, alpha_0, b, s, beta_0}),
followed by the parameters for the lifetime covariate at original scale (\code{mCov_life})
and then followed by the parameters for the transaction covariate at original scale
\code{mCov_trans}.\cr
\code{r, alpha_0, b, s, beta_0} are the log()-ed model parameters
used for estimation, in this order.\cr
\code{s}: shape parameter of the Gamma distribution for the lifetime process.
The smaller \code{s}, the stronger the heterogeneity of customer lifetimes.\cr
\code{beta}: scale parameter for the Gamma distribution for the lifetime process.\cr
\code{b:} scale parameter of the Gompertz distribution (constant across customers).\cr
\code{r:} shape parameter of the Gamma distribution of the purchase process.
The smaller \code{r}, the stronger the heterogeneity of the pruchase process.\cr
\code{alpha}: scale parameter of the Gamma distribution of the purchase process.\cr
\code{mCov_life}: parameters for the covariates affecting the lifetime process.\cr
\code{mCov_trans}: parameters for the covariates affecting the transaction process.

\code{mCov_trans} is a matrix containing the covariates data of
the time-invariant covariates that affect the transaction process.
Each column represents a different covariate. For every column a gamma parameter
needs to added to \code{vCovParams_trans} at the respective position.

\code{mCov_life} is a matrix containing the covariates data of
the time-invariant covariates that affect the lifetime process.
Each column represents a different covariate. For every column a gamma parameter
needs to added to \code{vCovParams_life} at the respective position.
}
