% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_interface_gnbd.R
\name{gnbd}
\alias{gnbd}
\alias{gnbd,clv.data-method}
\alias{gnbd,clv.data.static.covariates-method}
\alias{gnbd,clv.data.dynamic.covariates-method}
\title{GG/NBD models - Work In Progress}
\usage{
\S4method{gnbd}{clv.data}(
  clv.data,
  start.params.model = c(),
  optimx.args = list(),
  verbose = TRUE,
  ...
)

\S4method{gnbd}{clv.data.static.covariates}(
  clv.data,
  start.params.model = c(),
  optimx.args = list(),
  verbose = TRUE,
  names.cov.life = c(),
  names.cov.trans = c(),
  start.params.life = c(),
  start.params.trans = c(),
  names.cov.constr = c(),
  start.params.constr = c(),
  reg.lambdas = c(),
  ...
)

\S4method{gnbd}{clv.data.dynamic.covariates}(
  clv.data,
  start.params.model = c(),
  optimx.args = list(),
  verbose = TRUE,
  names.cov.life = c(),
  names.cov.trans = c(),
  start.params.life = c(),
  start.params.trans = c(),
  names.cov.constr = c(),
  start.params.constr = c(),
  reg.lambdas = c(),
  ...
)
}
\description{
Fits GG/NBD models on transactional data with static and without covariates.
}
