% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gsl_hyp_2F1}
\alias{gsl_hyp_2F1}
\title{GSL Hypergeom 2f1}
\usage{
gsl_hyp_2F1(a, b, c, x)
}
\value{
Returns the value of the gsl_sf_hyperg_2F1 call.
}
\description{
Calculate the hypergeometric 2f1 using the GSL library (gsl_sf_hyperg_2F1_e)
}
\keyword{internal}
