% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_interface_latentattrition.R
\name{latentAttrition}
\alias{latentAttrition}
\title{Formula Interface for Latent Attrition Models}
\usage{
latentAttrition(formula, data, cov, optimx.args = list(), verbose = TRUE)
}
\arguments{
\item{formula}{Formula specifying the model to be fit. See Details.}

\item{data}{Either a \code{clv.data} object or a \code{data.frame} containing transaction
data on which the model specified in \code{formula} will be fit.}

\item{cov}{Optional data.frame or data.table of covariate data.}

\item{optimx.args}{Additional arguments to control the optimization which are forwarded to \code{\link[optimx:optimx]{optimx::optimx}}.
If multiple optimization methods are specified, only the result of the last method is further processed.}

\item{verbose}{Show details about the running of the function.}
}
\description{
Fit latent attrition models for transaction with a formula interface
}
\details{
The formula argument follows a multipart notation:
A formula describing how to prepare and the model.
If data is provided as data.frame, the formula is required to have a LHS


See the example section for illustrations on how to specify the formula parameter.
}
\examples{
\donttest{

data("apparelTrans")
data("apparelStaticCov")

clv.nocov <-
    clvdata(apparelTrans, time.unit="w", date.format="ymd")

# Create static covariate data with 2 covariates
clv.staticcov  <-
  SetStaticCovariates(clv.nocov,
                      data.cov.life  = apparelStaticCov,
                      names.cov.life = c("Gender", "Channel"),
                      data.cov.trans = apparelStaticCov,
                      names.cov.trans = c("Gender", "Channel"))

# Fit pnbd without covariates
latentAttrition(~pnbd(), data=clv.nocov)
# Fit bgnbd without covariates
latentAttrition(~bgnbd(), data=clv.nocov)
# Fit ggomnbd without covariates
latentAttrition(~ggomnbd(), data=clv.nocov)

# Fit pnbd with start.parameters and correlation
latentAttrition(~pnbd(start.params.model=c(r=1, alpha=10, s=2, beta=8),
                      use.cor=TRUE),
                data=clv.nocov)

# Fit pnbd with all present covariates
latentAttrition(~pnbd()|.|., clv.staticcov)

# Fit pnbd with selected covariates
latentAttrition(~pnbd()|Gender|Channel+Gender, clv.staticcov)

# Fit pnbd with transformed covariate data
latentAttrition(~pnbd()|Gender|I(log(Channel+2)), clv.staticcov)

# Fit pnbd with all covs and regularization
latentAttrition(~pnbd()|.|.|regularization(life=3, trans=8), clv.staticcov)

# Fit pnbd on given data.frame, no split
latentAttrition(data()~pnbd(), data=apparelTrans)

# Fit pnbd, split data after 39 periods
latentAttrition(data(split=39)~pnbd(), data=apparelTrans)
# Same but also give date format and period definition
latentAttrition(data(split=39, format=ymd, unit=w)~pnbd(), data=apparelTrans)

# Fit pnbd on given data.frame w/ all covariates
latentAttrition(data()~pnbd()|.|., data=apparelTrans, cov=apparelStaticCov)

# Fit pnbd on given data.frame w/ selected covariates
latentAttrition(data()~pnbd()|Channel+Gender|Gender,
                data=apparelTrans, cov=apparelStaticCov)

}


}
\seealso{
Models for inputs to: \link{pnbd}, \link{ggomnbd}, \link{bgnbd}.

\link{spending} to fit spending models with a formula interface
}
