% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_s3generics_clvfitted.R
\name{nobs.clv.fitted}
\alias{nobs.clv.fitted}
\title{Number of observations}
\usage{
\method{nobs}{clv.fitted}(object, ...)
}
\arguments{
\item{object}{An object of class clv.fitted.}

\item{...}{Ignored}
}
\value{
The number of customers.
}
\description{
The number of observations is defined as the number of unique customers for which the model was fit.
}
