% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_s3generics_clvfittedspending_plot.R
\name{plot.clv.fitted.spending}
\alias{plot.clv.fitted.spending}
\alias{plot,clv.fitted.spending-method}
\title{Plot expected and actual spending density}
\usage{
\method{plot}{clv.fitted.spending}(x, verbose = TRUE, plot.interpolation.points = 250, ...)

\S4method{plot}{clv.fitted.spending}(x, verbose = TRUE, plot.interpolation.points = 250, ...)
}
\arguments{
\item{x}{The fitted clv model to plot}

\item{verbose}{Show details about the running of the function.}

\item{plot.interpolation.points}{Number of interpolation points in density graph}

\item{...}{Ignored}
}
\value{
An object of class \code{ggplot} from package \code{ggplot2} is returned by default.
}
\description{
Plot the spending density (actual vs. model-based)
}
\examples{
\donttest{

library("CLVTools")
library("ggplot2")

data("cdnow")

clv.cdnow <- clvdata(cdnow,
  date.format="ymd",
  time.unit = "week",
  estimation.split = "1997-09-30",
  name.id = "Id",
  name.date = "Date",
  name.price = "Price")

est.gg <- gg(clv.data = clv.cdnow, remove.first.transaction = FALSE)

plot(est.gg, plot.interpolation.points = 1000)

}

}
