% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_interface_pmf.R
\name{pmf}
\alias{pmf}
\alias{pmf,clv.fitted.transactions-method}
\title{Probability Mass Function}
\usage{
\S4method{pmf}{clv.fitted.transactions}(object, x = 0:5)
}
\arguments{
\item{object}{Fitted transaction model xxxx}

\item{x}{Integer vector}
}
\description{
Calculate P(X(t)=x), the probability to make exactly x transactions in the interval [0, t) (xxxx or ]??).
}
