% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pnbd_dyncov_pmf_A_i_C}
\alias{pnbd_dyncov_pmf_A_i_C}
\title{Get transaction covariate effect for a specific time period}
\arguments{
\item{i_1based}{One-based index of the period to retrieve}

\item{dt_data_period_customer_trans}{Dynamic covariates object containing transaction effects}
}
\value{
Covariate effect value for the specified period
}
\description{
Retrieves the transaction covariate effect for a specific period index
}
\keyword{internal}
