% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pnbd_dyncov_pmf_Bbar_i_C}
\alias{pnbd_dyncov_pmf_Bbar_i_C}
\title{Calculate adjusted cumulative transaction covariate effect}
\arguments{
\item{i_1based}{One-based index of the period to calculate up to}

\item{dt_data_period_customer_trans}{Dynamic covariates object containing transaction effects}

\item{d1}{Fraction of time between first transaction and next covariate date}

\item{ui}{Time between customer's first transaction and start of estimation period}
}
\value{
Adjusted cumulative transaction covariate effect
}
\description{
Computes Bbar_i, the adjusted cumulative transaction covariate effect
}
\keyword{internal}
