% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pnbd_dyncov_pmf_bu_i_C}
\alias{pnbd_dyncov_pmf_bu_i_C}
\title{Calculate time boundary for period i}
\arguments{
\item{ui}{Time between customer's first transaction and start of estimation period}

\item{i_1based}{One-based index of the period}

\item{d1}{Fraction of time between first transaction and next covariate date}
}
\value{
Time boundary value
}
\description{
Computes the time boundary bu_i for the specified period
}
\keyword{internal}
