% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pnbd_dyncov_pmf_Dbar_i_C}
\alias{pnbd_dyncov_pmf_Dbar_i_C}
\title{Calculate adjusted cumulative dropout covariate effect}
\arguments{
\item{i_1based_period}{One-based index of the period to calculate}

\item{dt_data_period_customer_life}{Dynamic covariates object for lifetime in period}

\item{dt_data_since_alive_customer_life}{Dynamic covariates object for lifetime since alive}

\item{d_omega}{Fraction of time from period start until next time unit}

\item{k0u_Dbar}{Number of time units between customer's first transaction and start of period}
}
\value{
Adjusted cumulative dropout covariate effect
}
\description{
Computes Dbar_i, the adjusted cumulative dropout covariate effect
}
\keyword{internal}
