% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pnbd_dyncov_pmf_S1_per_customer_C}
\alias{pnbd_dyncov_pmf_S1_per_customer_C}
\title{Calculate S1 component of PMF}
\arguments{
\item{dt_data_period_customer_trans}{Dynamic covariates for transaction process}

\item{dt_data_period_customer_life}{Dynamic covariates for lifetime process}

\item{dt_data_since_alive_customer_life}{Dynamic covariates for lifetime since first transaction}

\item{x_double}{Number of transactions}

\item{alpha_r}{Scale parameter for transaction rate distribution}

\item{beta_s}{Scale parameter for dropout rate distribution}

\item{r}{Shape parameter for transaction rate distribution}

\item{s}{Shape parameter for dropout rate distribution}

\item{t_r}{Time span for prediction}

\item{ui}{Time between customer's first transaction and start of estimation period}

\item{d1}{Fraction of time between first transaction and next covariate date}

\item{d_omega}{Fraction of time from period start until next time unit}

\item{k0u_S1}{Number of time units between customer's first transaction and period start}
}
\value{
S1 component value
}
\description{
Computes the S1 component of the Pareto/NBD PMF with dynamic covariates
}
\details{
S1 represents the probability that a customer makes exactly x transactions in the 
prediction period conditional on being alive throughout the entire period.
}
\keyword{internal}
