% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pnbd_dyncov_pmf_S2_1j_per_customer_C}
\alias{pnbd_dyncov_pmf_S2_1j_per_customer_C}
\title{Calculate S2_1j component of PMF}
\arguments{
\item{dt_data_period_customer_trans}{Dynamic covariates for transaction process}

\item{dt_data_period_customer_life}{Dynamic covariates for lifetime process}

\item{dt_data_since_alive_customer_life}{Dynamic covariates for lifetime since first transaction}

\item{j_S2_1j}{Number of transactions at time of churn}

\item{x_double}{Total number of transactions}

\item{alpha_r}{Scale parameter for transaction rate distribution}

\item{beta_s}{Scale parameter for dropout rate distribution}

\item{r_param}{Shape parameter for transaction rate distribution}

\item{s_param}{Shape parameter for dropout rate distribution}

\item{ui}{Time between customer's first transaction and start of estimation period}

\item{d1}{Fraction of time between first transaction and next covariate date}

\item{d_omega}{Fraction of time from period start until next time unit}

\item{k0u_S2_1j}{Number of time units between customer's first transaction and period start}
}
\value{
S2_1j component value
}
\description{
Computes the S2_1j component of the Pareto/NBD PMF with dynamic covariates
}
\details{
S2_1j represents the probability that a customer churns during the first period
with exactly j transactions at the time of churn, and makes a total of x transactions.
}
\keyword{internal}
