% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pnbd_nocov_CET}
\alias{pnbd_nocov_CET}
\title{Pareto/NBD: Conditional Expected Transactions without covariates}
\usage{
pnbd_nocov_CET(vEstimated_params, dPrediction_period, vX, vT_x, vT_cal)
}
\arguments{
\item{vEstimated_params}{Vector with the Pareto/NBD model parameters at original scale.}

\item{dPrediction_period}{time prediction time frame}

\item{vX}{Frequency vector of length n counting the numbers of purchases.}

\item{vT_x}{Recency vector of length n.}

\item{vT_cal}{Vector of length n indicating the total number of periods of observation.}
}
\value{
Returns a vector containing the conditional expected transactions for the existing
customers in the Pareto/NBD model.
}
\description{
Calculates the expected number of transactions in a given time period based
on a customer's past transaction behavior and the Pareto/NBD model parameters.
}
\details{
\code{vEstimated_params} vector with the estimated parameters in original scale
for the Pareto/NBD model, namely (r, alpha, s, beta).
r and alpha: unobserved parameters that describe the NBD transaction process.
s and beta: unobserved parameters that describe the Pareto
(exponential gamma) dropout process.
}
\references{
Fader, Peter S., and Bruce G.S. Hardie (2005). "A Note on Deriving the
 Pareto/NBD Model and Related Expressions.", Web.
 \url{http://www.brucehardie.com/notes/008/}.
}
