% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pnbd_nocov_LL_ind}
\alias{pnbd_nocov_LL_ind}
\alias{pnbd_nocov_LL_sum}
\title{Pareto/NBD: LogLikelihood without covariates}
\usage{
pnbd_nocov_LL_ind(vLogparams, vX, vT_x, vT_cal)

pnbd_nocov_LL_sum(vLogparams, vX, vT_x, vT_cal)
}
\arguments{
\item{vLogparams}{vector with the Pareto/NBD model parameters log scaled}

\item{vX}{Frequency vector of length n counting the numbers of purchases.}

\item{vT_x}{Recency vector of length n.}

\item{vT_cal}{Vector of length n indicating the total number of periods of observation.}
}
\value{
Returns the respective LogLikelihood value for the Pareto/NBD model without covariates.
}
\description{
Pareto/NBD without Covariates:

The function \code{pnbd_nocov_LL_ind} calculates the individual LogLikelihood
values for each customer for the given parameters.

The function \code{pnbd_nocov_LL_sum} calculates the LogLikelihood value summed
across customers for the given parameters.
}
\details{
\code{r, alpha_0, s, beta_0} are the parameters used for estimation.\cr
s: shape parameter of the Gamma distribution for the lifetime process.
The smaller s, the stronger the heterogeneity of customer lifetimes. \cr
beta: scale parameter for the Gamma distribution for the lifetime process. \cr
r: shape parameter of the Gamma distribution of the purchase process.
The smaller r, the stronger the heterogeneity of the purchase process.\cr
alpha: scale parameter of the Gamma distribution of the purchase process.
}
\references{
Fader, Peter S., and Bruce G.S. Hardie (2005). "A Note on Deriving the
 Pareto/NBD Model and Related Expressions.", Web.
 \url{http://www.brucehardie.com/notes/008/}.
}
