% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pnbd_nocov_PAlive}
\alias{pnbd_nocov_PAlive}
\title{Pareto/NBD: PAlive without covariates}
\usage{
pnbd_nocov_PAlive(vEstimated_params, vX, vT_x, vT_cal)
}
\arguments{
\item{vEstimated_params}{Vector with the Pareto/NBD model parameters at original scale.}

\item{vX}{Frequency vector of length n counting the numbers of purchases.}

\item{vT_x}{Recency vector of length n.}

\item{vT_cal}{Vector of length n indicating the total number of periods of observation.}
}
\value{
Returns a vector with the PAlive for each customer.
}
\description{
Pareto/NBD without Covariates: Calculates the probability of a customer being alive
at the end of the calibration period.
}
\details{
\code{vEstimated_params} vector with the estimated parameters in original scale
for the Pareto/NBD model, namely (r, alpha, s, beta).
r and alpha: unobserved parameters that describe the NBD transaction process.
s and beta: unobserved parameters that describe the pareto
(exponential gamma) dropout process.
}
\references{
Fader, Peter S., and Bruce G.S. Hardie (2005). "A Note on Deriving the
 Pareto/NBD Model and Related Expressions.", Web.
 \url{http://www.brucehardie.com/notes/008/}.
}
