% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_interface_predictnewcustomers.R
\name{predict.new.customers}
\alias{predict.new.customers}
\alias{predict.new.customers,clv.fitted.transactions-method}
\alias{predict.new.customers,clv.fitted.transactions.static.cov-method}
\alias{predict.new.customers,clv.fitted.transactions.dynamic.cov-method}
\title{Expected number of transactions of an average new customers}
\usage{
\S4method{predict.new.customers}{clv.fitted.transactions}(clv.fitted, t, ...)

\S4method{predict.new.customers}{clv.fitted.transactions.static.cov}(clv.fitted, t, data.cov.life, data.cov.trans, ...)

\S4method{predict.new.customers}{clv.fitted.transactions.dynamic.cov}(
  clv.fitted,
  t,
  data.cov.life,
  data.cov.trans,
  first.transaction,
  ...
)
}
\arguments{
\item{clv.fitted}{A fitted transaction model for which prediction is desired.}

\item{t}{A positive, numeric scalar indicating the number of periods to predict.}

\item{...}{Ignored}

\item{data.cov.life}{Covariate data for the lifetime process for a single customer, \code{data.table} or \code{data.frame}. See details.}

\item{data.cov.trans}{Covariate data for the lifetime process for a single customer, \code{data.table} or \code{data.frame}. See details.}

\item{first.transaction}{For dynamic covariate models only: The time point of the first transaction.
Has to be within the time range of the covariate data.}
}
\value{
A numeric scalar with the expected number of transactions.
}
\description{
Predict the number of transactions a single, average customer is expected to make in
the \code{t} periods since making the first transaction ("coming alive").
For covariate models, the prediction is for an average customer with the given covariates.

The unconditional expectation \code{XXXXXFORMULAXXXX} is used for making this prediction.
For models without covariates, the prediction hence is the same for all customers
and independent of when a customer comes alive.
For models with covariates, the prediction is the same for all customers
with the same covariates.
For models with dynamic covariates, the time point of the first transaction is additionally
required because the exact covariates active during the prediction period have to be known.
}
\details{
The covariate data has to contain the same columns as during model fitting. Only numeric values are allowed.
No customer Id is required because no data of the customers used for model fitting is used for making this prediction.

The covariate data has to be of the following structure:
 Static: exactly 1 row of covariate data
 Dyn: exactly 1 row of covariate data for each covariate
For dynamic covariate models, the covariate data requires the following structure:

The covariate data has to cover the time from the customer's first transaction \code{first.transaction}
to the end of the prediction period given by \code{t}.
}
