% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_interface_spending.R
\name{spending}
\alias{spending}
\title{Formula Interface for Spending Models}
\usage{
spending(formula, data, optimx.args = list(), verbose = TRUE)
}
\description{
Formula Interface for Spending Models
}
