% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vec_gsl_hyp2f0_e}
\alias{vec_gsl_hyp2f0_e}
\title{GSL Hypergeom 2f0 for equal length vectors}
\usage{
vec_gsl_hyp2f0_e(vA, vB, vZ)
}
\arguments{
\item{vA}{Vector of values for parameter a}

\item{vB}{Vector of values for parameter b}

\item{vZ}{Vector of values for parameter z}
}
\value{
List with vector of values and vector of gsl status codes
}
\description{
Calculate the hypergeometric 2f0 using the GSL library (gsl_sf_hyperg_2F0_e)
}
\keyword{internal}
