% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{apparelDynCov}
\alias{apparelDynCov}
\title{Time-varying Covariates for the Apparel Retailer Dataset}
\format{
A data.table with 187,800 rows and 5 variables
\describe{
  \item{Id}{Customer Id}
  \item{Cov.Date}{Date of contextual factor}
  \item{High.Season}{Seasonal variable: 1 indicating a time-period that is considered "high season".}
  \item{Gender}{0=male, 1=female}
  \item{Channel}{Acquisition channel: 0=online, 1=offline}
}
}
\usage{
data("apparelDynCov")
}
\description{
This simulated data contains seasonal information and additional covariates on all 400 customers in the "apparelTrans" dataset.
This information can be used as time-varying covariates.
}
\keyword{datasets}
