% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{apparelTrans}
\alias{apparelTrans}
\title{Apparel Retailer Dataset}
\format{
A \code{data.table} with 3,187 rows and 3 variables:
\describe{
  \item{\code{Id}}{Customer Id}
  \item{\code{Date}}{Date of purchase}
  \item{\code{Price}}{Price of purchase}
}
}
\usage{
data("apparelTrans")
}
\description{
This is a simulated dataset containing the entire purchase history of customers made their first purchase at an
apparel retailer on January 2nd 2005. In total the dataset contains 600 customers who made
2,647 transactions between January 2005 and end of December 2010.
}
\keyword{datasets}
