% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{apparelTrans}
\alias{apparelTrans}
\title{Apparel Retailer Dataset}
\format{
A \code{data.table} with 19573 rows and 3 variables:
\describe{
  \item{\code{Id}}{Customer Id}
  \item{\code{Date}}{Date of purchase}
  \item{\code{Price}}{Price of purchase}
}
}
\usage{
data("apparelTrans")
}
\description{
A dataset containing the entire purchase history of customers made their first purchase at an
apparel retailer during January to May 2010. In total the dataset contains 2201 customers who made
19573 transactions between January 2010 and October 2016.
}
\keyword{datasets}
