% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_interface_bgnbd.R
\name{bgnbd}
\alias{bgnbd}
\alias{bgnbd,clv.data-method}
\alias{bgnbd,clv.data.static.covariates-method}
\alias{bgnbd,clv.data.dynamic.covariates-method}
\title{BG/NBD models}
\usage{
\S4method{bgnbd}{clv.data}(
  clv.data,
  start.params.model = c(),
  use.cor = FALSE,
  start.param.cor = c(),
  optimx.args = list(),
  verbose = TRUE,
  ...
)

\S4method{bgnbd}{clv.data.static.covariates}(
  clv.data,
  start.params.model = c(),
  use.cor = FALSE,
  start.param.cor = c(),
  optimx.args = list(),
  verbose = TRUE,
  names.cov.life = c(),
  names.cov.trans = c(),
  start.params.life = c(),
  start.params.trans = c(),
  names.cov.constr = c(),
  start.params.constr = c(),
  reg.lambdas = c(),
  ...
)

\S4method{bgnbd}{clv.data.dynamic.covariates}(
  clv.data,
  start.params.model = c(),
  optimx.args = list(),
  verbose = TRUE,
  names.cov.life = c(),
  names.cov.trans = c(),
  start.params.life = c(),
  start.params.trans = c(),
  names.cov.constr = c(),
  start.params.constr = c(),
  reg.lambdas = c(),
  ...
)
}
\arguments{
\item{clv.data}{The data object on which the model is fitted.}

\item{start.params.model}{Named start parameters containing the optimization start parameters for the model without covariates.}

\item{optimx.args}{Additional arguments to control the optimization which are forwarded to \code{\link[optimx:optimx]{optimx::optimx}}.
If multiple optimization methods are specified, only the result of the last method is further processed.}

\item{verbose}{Show details about the running of the function.}

\item{...}{Ignored}

\item{names.cov.life}{Which of the set Lifetime covariates should be used. Missing parameter indicates all covariates shall be used.}

\item{names.cov.trans}{Which of the set Transaction covariates should be used. Missing parameter indicates all covariates shall be used.}

\item{start.params.life}{Named start parameters containing the optimization start parameters for all lifetime covariates.}

\item{start.params.trans}{Named start parameters containing the optimization start parameters for all transaction covariates.}

\item{names.cov.constr}{Which covariates should be forced to use the same parameters for the lifetime and transaction process. The covariates need to be present as both, lifetime and transaction covariates.}

\item{start.params.constr}{Named start parameters containing the optimization start parameters for the constraint covariates.}

\item{reg.lambdas}{Named lambda parameters used for the L2 regularization of the lifetime and the transaction covariate parameters. Lambdas have to be >= 0.}
}
\value{
\code{bgnbd} returns an object of
class \code{clv.bgnbd}.

The function \code{\link[CLVTools:summary.clv.fitted]{summary}} can be used to obtain and print a summary of the results.
The generic accessor functions \code{coefficients}, \code{fitted},
\code{residuals}, \code{vcov}, \code{logLik}, \code{AIC}, \code{BIC}, and \code{nobs} are available.
}
\description{
Fits BG/NBD models on transactional data without covariates.
}
\details{
Model parameters for the BG/NBD model are \code{r, alpha, a, and b}. \cr
\code{r}: TODO \cr
\code{alpha}: TODO \cr
\code{a}: TODO \cr
\code{b}: TODO

If no start parameters are given, r = 1, alpha = 3, a = 1, b = 3 is used.
The model start parameters are required to be > 0.

\subsection{The BG/NBD model}{
TODO
}
}
\examples{
\donttest{
data("cdnow")
clv.data.cdnow <- clvdata(cdnow, date.format = "ymd",
                            time.unit = "w", estimation.split = 37)

# Fit standard BG/NBD model
bgnbd(clv.data.cdnow)

# Give initial guesses for the Model parameters
bgnbd(clv.data.apparel,
     start.params.model = c(r=1.1, alpha=3.4, a=1, beta=3))


# pass additional parameters to the optimizer (optimx)
#    Use Nelder-Mead as optimization method and print
#    detailed information about the optimization process
estimation.bgnbd <- bgnbd(clv.data.cdnow,
                     optimx.args = list(method="Nelder-Mead",
                                        control=list(trace=6)))

# estimated coefs
coef(estimation.bgnbd)

# summary of the fitted model
summary(estimation.bgnbd)

predict.bgnbd <- predict(estimation.bgnbd, prediction.end = "2011-12-31")

plot(estimation.bgnbd)
}
}
\references{
BG/NBD model according to
\url{https://github.com/cran/BTYD}
}
\seealso{
\code{\link[CLVTools:clvdata]{clvdata}} to create a clv data object

\code{\link[CLVTools:predict.clv.fitted]{predict}} to predict expected transactions, probability of being alive, and customer lifetime value for every customer

\code{\link[CLVTools:plot.clv.fitted]{plot}} to plot the unconditional expectation as predicted by the fitted model

The generic functions \code{\link[CLVTools:summary.clv.fitted]{summary}} and \code{\link[CLVTools:fitted.clv.fitted]{fitted}}.
}
