% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bgnbd_CET}
\alias{bgnbd_CET}
\alias{bgnbd_nocov_CET}
\title{BG/BB: Conditional Expected Transactions without covariates}
\usage{
bgnbd_nocov_CET(
  alpha,
  beta,
  gamma,
  delta,
  nPeriods,
  vX,
  vT_x,
  vT_cal,
  vN_cal,
  vN_star
)
}
\arguments{
\item{alpha}{TODO: describe}

\item{beta}{TODO: describe}

\item{gamma}{TODO: describe}

\item{delta}{TODO: describe}

\item{nPeriods}{time prediction time frame}

\item{vX}{Frequency vector of length n counting the numbers of purchases.}

\item{vT_x}{Recency vector of length n.}

\item{vT_cal}{Vector of length n indicating the total number of periods of observation.}
}
\value{
Returns a vector containing the conditional expected transactions for the existing
customers in the BG/BB model.
}
\description{
Calculates the expected number of transactions in a given time period based
on a customer's past transaction behavior and the BG/NBD model parameters.
}
\details{
\code{alpha} \code{beta} \code{gamma} \code{delta} estimated parameters in original scale
for the BG/BB model, namely (alpha, beta, gamma, delta).
}
