% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bgnbd_PAlive}
\alias{bgnbd_PAlive}
\alias{bgnbd_palive}
\title{BG/NBD: PAlive without covariates}
\usage{
bgnbd_palive(vParams, vX, vT_x, vT_cal)
}
\arguments{
\item{vX}{Frequency vector of length n counting the numbers of purchases.}

\item{vT_x}{Recency vector of length n.}

\item{vT_cal}{Vector of length n indicating the total number of periods of observation.}

\item{vEstimated_params}{Vector with the Pareto/NBD model parameters at original scale.}
}
\value{
Returns a vector with the PAlive for each customer.
}
\description{
BG/NBD without Covariates: Calculates the probability of a customer being alive
at the end of the calibration period.
}
\details{
\code{vParams} vector with the estimated parameters in original scale
for the BG/NBD model, namely (r, alpha, a, b).
r and alpha: TODO: description.
a and b: TODO: description
}
