% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_clv_fitted_transactions.R
\docType{class}
\name{clv.fitted.transactions-class}
\alias{clv.fitted.transactions-class}
\title{Fitted Transaction Model without covariates}
\description{
Extends the class \code{clv.fitted} to performs steps during the
estimation, prediction and plotting process that are specific to all transaction models.
}
\seealso{
Definition of the parent class \linkS4class{clv.fitted}

For spending models \linkS4class{clv.fitted.spending}
}
\keyword{internal}
