% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_clv_gg.R
\docType{class}
\name{clv.gg-class}
\alias{clv.gg-class}
\title{Result of fitting the Gamma-Gamma model}
\description{
Output from fitting the Gamma-Gamma model on transactional data. It constitutes the estimation
result and is returned to the user to use it as input to other methods such as to diagnose the model
fit or to predict customer's future spending.

Inherits from \linkS4class{clv.fitted.spending} in order to execute all steps required to fit a model or to predict
future spending. Contains an instance of class \linkS4class{clv.model.gg} which
provides the required Gamma-Gamma specific functionalities.
}
\section{Slots}{

\describe{
\item{\code{cbs}}{Single \code{data.table} that contains information about
each customers' mean spending per transaction (Spending) and the number of purchases (x) which
both depend on whether the first transaction was removed or not.}
}}

\seealso{
\linkS4class{clv.fitted}, \linkS4class{clv.fitted.spending}, \linkS4class{clv.model.gg}
}
\keyword{internal}
