% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_clv_model_bgnbd.R
\docType{class}
\name{clv.model.bgnbd.no.cov-class}
\alias{clv.model.bgnbd.no.cov-class}
\title{CLV Model functionality for BG/NBD without covariates}
\description{
This class implements the functionalities and model-specific steps which are required
to fit the BG/NBD model without covariates.
}
\seealso{
Other clv model classes \linkS4class{clv.model}, \linkS4class{clv.model.bgnbd.static.cov}

Classes using its instance: \linkS4class{clv.fitted}
}
\keyword{internal}
