% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_clv_model_gg.R
\docType{class}
\name{clv.model.gg-class}
\alias{clv.model.gg-class}
\title{CLV Model functionality for the Gamma-Gamma spending model}
\description{
This class implements the functionalities and model-specific steps which are required
to fit the Gamma-Gamma spending model.
}
\keyword{internal}
