% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_clv_model_pnbd_staticcov.R
\docType{class}
\name{clv.model.pnbd.static.cov-class}
\alias{clv.model.pnbd.static.cov-class}
\title{CLV Model functionality for Pareto/NBD with static covariates}
\description{
This class implements the functionalities and model-specific steps which are required
to fit the Pareto/NBD model with static covariates.
}
\seealso{
Other clv model classes \linkS4class{clv.model}, \linkS4class{clv.model.pnbd.no.cov}, \linkS4class{clv.model.pnbd.dynamic.cov}

Classes using its instance: \linkS4class{clv.fitted.transactions.static.cov}
}
\keyword{internal}
