% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_clv_time_days.R
\docType{class}
\name{clv.time.days-class}
\alias{clv.time.days-class}
\title{Time unit representing a single Day}
\description{
Inherits from \code{clv.time.date} and therefore stores all timepoints
with instances of class \code{Date}. Hence, no temporal resolution finer than 1 day is
possible with this time unit.
}
\seealso{
\linkS4class{clv.time}

\linkS4class{clv.time.date}
}
\keyword{internal}
