% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_interface_hessian.R
\name{hessian}
\alias{hessian}
\alias{hessian.clv.fitted}
\alias{hessian,clv.fitted-method}
\title{Calculate hessian for a fitted model}
\usage{
\method{hessian}{clv.fitted}(object, method.args = list())

hessian(object, ...)

\S4method{hessian}{clv.fitted}(object, method.args = list())
}
\arguments{
\item{object}{Fitted model}

\item{method.args}{List of options forwarded to the numerical approximation
method. See \link[numDeriv:hessian]{numDeriv::hessian}.}

\item{...}{Ignored}
}
\value{
The hessian matrix, with column and row names set to the parameter
names used to call the LL.
}
\description{
Calculate a numerical approximation to the Hessian matrix at
the final estimated parameters using \code{numDeriv::hessian}.
}
