% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_interface_latentattrition.R
\name{latentAttrition}
\alias{latentAttrition}
\title{Formula Interface for Transaction Models}
\usage{
latentAttrition(formula, data, optimx.args = list(), verbose = TRUE)
}
\description{
Formula Interface for Transaction Models
}
