% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_s3generics_clvdata.R
\name{nobs.clv.data}
\alias{nobs.clv.data}
\title{Number of observations}
\usage{
\method{nobs}{clv.data}(object, ...)
}
\arguments{
\item{object}{An object of class clv.data.}

\item{...}{Ignored}
}
\value{
The number of customers.
}
\description{
The number of observations is defined as the number of unique customers in the transaction data.
}
