% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_interface_pmf.R
\name{pmf}
\alias{pmf}
\alias{pmf,clv.fitted.transactions-method}
\title{Probability Mass Function}
\usage{
\S4method{pmf}{clv.fitted.transactions}(object, x = 0:5)
}
\arguments{
\item{object}{The fitted transaction model.}

\item{x}{Vector of positive integer numbers (>=1) indicating the number of transactions x for
which the PMF should be calculated.}
}
\value{
Returns a \code{data.table} with ids and depending on \code{x}, multiple columns of PMF values, each column
for one value in \code{x}.
\item{Id}{customer identification}
\item{pmf.x.Y}{PMF values for Y number of transactions}
}
\description{
Calculate P(X(t)=x), the probability to make exactly x repeat transactions in
the interval (0, T] (the estimation period). Note that T is T.cal, the observation period which
differs by customer.
}
\examples{
\donttest{
data("cdnow")

# Fit the ParetoNBD model on the CDnow data
pnbd.cdnow <- pnbd(clvdata(cdnow, time.unit="w",
                           estimation.split=37,
                           date.format="ymd"))

# Calculate the PMF for 0 to 10 transactions
#  in the estimation period
pmf(pnbd.cdnow, x=0:10)

# Compare vs. actuals (CBS):
*** TODO: NO, pmf is per customer!!
x    pmf             actual percentage of x
1    0.163130227     439/2357 = 0.18625371
2    0.079697124     214/2357 = 0.09079338
3    0.046111155     100/2357 = 0.04242681
4    0.028749294     62/2357  = 0.02630462
5    0.018591868     38/2357  = 0.01612219
6    0.012213620     29/2357  = 0.01230378
7    0.008029474     23/2357  = 0.00975817
8    0.005206442     7/2357   = 0.00296988
9    0.003266914     5/2357   = 0.00212134
10   0.001919501     5/2357   = 0.00212134
}

}
\seealso{
The model fitting functions \code{\link[CLVTools:pnbd]{pnbd},
\link[CLVTools:bgnbd]{bgnbd}, \link[CLVTools:ggomnbd]{ggomnbd}}.
}
